Implement a provider-model configurable at CLI
