+++ internal/mysql/tables_test.go
@@ -8,6 +8,7 @@ import (
 	"testing"
 
 	"github.com/DATA-DOG/go-sqlmock"
+	"github.com/skpr/mtk/internal/mysql/providers"
 	"github.com/stretchr/testify/assert"
 
 	"github.com/skpr/mtk/internal/mysql/mock"
@@ -97,88 +98,12 @@ func TestMySQLDumpCreateTableHandlingErrorWhenScanningRows(t *testing.T) {
 	assert.NotNil(t, dumper.WriteCreateTable(buffer, "table"))
 }
 
-func TestMySQLGetColumnsForSelect(t *testing.T) {
-	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
-	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnRows(
-		sqlmock.NewRows([]string{"col1", "col2", "col3"}).AddRow("a", "b", "c"))
-	columns, err := dumper.QueryColumnsForTable("table", DumpParams{
-		SelectMap: map[string]map[string]string{"table": {"col2": "NOW()"}},
-	})
-	assert.Nil(t, err)
-	assert.Equal(t, []string{"`col1`", "NOW() AS `col2`", "`col3`"}, columns)
-}
-
-func TestMySQLGetColumnsForSelectHandlingErrorWhenQuerying(t *testing.T) {
-	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
-	error := errors.New("broken")
-	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnError(error)
-	columns, err := dumper.QueryColumnsForTable("table", DumpParams{
-		SelectMap: map[string]map[string]string{"table": {"col2": "NOW()"}},
-	})
-	assert.Equal(t, err, error)
-	assert.Empty(t, columns)
-}
-
-func TestMySQLGetSelectQueryFor(t *testing.T) {
-	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
-	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnRows(
-		sqlmock.NewRows([]string{"c1", "c2"}).AddRow("a", "b"))
-	query, err := dumper.GetSelectQueryForTable("table", DumpParams{
-		SelectMap: map[string]map[string]string{"table": {"c2": "NOW()"}},
-		WhereMap:  map[string]string{"table": "c1 > 0"},
-	})
-	assert.Nil(t, err)
-	assert.Equal(t, "SELECT `c1`, NOW() AS `c2` FROM `table` WHERE c1 > 0", query)
-}
-
-func TestMySQLGetExportSelectQueryFor(t *testing.T) {
-	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
-	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnRows(
-		sqlmock.NewRows([]string{"c1", "c2"}).AddRow("a", "b"))
-	query, err := dumper.GetSelectIntoOutFileQueryForTable("table", DumpParams{
-		SelectMap:  map[string]map[string]string{"table": {"c2": "NOW()"}},
-		WhereMap:   map[string]string{"table": "c1 > 0"},
-		DataExport: true,
-		DataPath:   "s3://path/to/bucket",
-		Region:     "ap-southeast-4",
-	})
-	assert.Nil(t, err)
-	assert.Equal(t, "SELECT `c1`, NOW() AS `c2` FROM `table` WHERE c1 > 0 INTO OUTFILE S3 's3://path/to/bucket/table.csv' FIELDS TERMINATED BY ',' ENCLOSED BY '\"' LINES TERMINATED BY '\\n' MANIFEST ON OVERWRITE ON", query)
-
-}
-
-func TestMySQLGetImportSelectQueryFor(t *testing.T) {
-	db, _ := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
-	query, err := dumper.GetLoadDataQueryForTable("table_name", "s3://path/to/bucket", "ap-southeast-4")
-	assert.Nil(t, err)
-	assert.Equal(t, "LOAD DATA FROM S3 FILE 'S3-ap-southeast-4://path/to/bucket/table_name.csv.manifest' INTO TABLE `table_name` FIELDS TERMINATED BY ',' ENCLOSED BY '\"' LINES TERMINATED BY '\\n'", query)
-
-}
-
-func TestMySQLGetSelectQueryForHandlingError(t *testing.T) {
-	db, mock := mock.GetDB(t)
-	dumper := NewClient(db, log.New(os.Stdout, "", 0))
-	error := errors.New("broken")
-	mock.ExpectQuery("SELECT \\* FROM `table` LIMIT 1").WillReturnError(error)
-	query, err := dumper.GetSelectQueryForTable("table", DumpParams{
-		SelectMap: map[string]map[string]string{"table": {"c2": "NOW()"}},
-		WhereMap:  map[string]string{"table": "c1 > 0"},
-	})
-	assert.Equal(t, error, err)
-	assert.Equal(t, "", query)
-}
-
 func TestMySQLGetRowCount(t *testing.T) {
 	db, mock := mock.GetDB(t)
 	dumper := NewClient(db, log.New(os.Stdout, "", 0))
 	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table` WHERE c1 > 0").WillReturnRows(
 		sqlmock.NewRows([]string{"COUNT(*)"}).AddRow(1234))
-	count, err := dumper.GetRowCountForTable("table", DumpParams{
+	count, err := dumper.GetRowCountForTable("table", providers.DumpParams{
 		WhereMap: map[string]string{"table": "c1 > 0"},
 	})
 	assert.Nil(t, err)
@@ -190,7 +115,7 @@ func TestMySQLGetRowCountHandlingError(t *testing.T) {
 	dumper := NewClient(db, log.New(os.Stdout, "", 0))
 	mock.ExpectQuery("SELECT COUNT\\(\\*\\) FROM `table` WHERE c1 > 0").WillReturnRows(
 		sqlmock.NewRows([]string{"COUNT(*)"}).AddRow(nil))
-	count, err := dumper.GetRowCountForTable("table", DumpParams{
+	count, err := dumper.GetRowCountForTable("table", providers.DumpParams{
 		WhereMap: map[string]string{"table": "c1 > 0"},
 	})
 	assert.NotNil(t, err)
